% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addCranInfo}
\alias{addCranInfo}
\title{Add CRAN badge info to README}
\usage{
addCranInfo(pkg = as.list(read.dcf("DESCRIPTION")[1, ])$Package)
}
\arguments{
\item{pkg}{\code{\link[base]{character}}.
Package name.}
}
\description{
TODO
}
\examples{
## TODO
}

