% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addDevtoolsComponents}
\alias{addDevtoolsComponents}
\title{Add components provided by devtools}
\usage{
addDevtoolsComponents()
}
\arguments{
\item{inst}{A class instance.}
}
\value{
An error. See respective methods.
}
\description{
Calls certain functions of \code{\link{devtools}}.
}
\details{
\itemize{
 \item{\code{\link[devtools]{use_testthat}}}
 \item{\code{\link[devtools]{use_data}}}
 \item{\code{\link[devtools]{use_build_ignore}}}
 \item{\code{\link[devtools]{use_travis}}}
 \item{\code{\link[devtools]{use_coverage}}}
 \item{\code{\link[devtools]{use_revdep}}}
 \item{\code{\link[devtools]{use_package_doc}}}
 \item{\code{\link[devtools]{use_vignette}}}
 \item{\code{\link[devtools]{use_cran_badge}}}
}
}
\examples{
## TODO
}

