% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addExample}
\alias{addExample}
\title{Add example}
\usage{
addExample(name = character())
}
\arguments{
\item{name}{\code{\link[base]{character}}.
Name for example file.}
}
\value{
\code{\link[base]{character}}. Path to example file.
}
\description{
Add example.
}
\details{
TODO
}
\examples{
## TODO
}

