% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addExamples}
\alias{addExamples}
\title{Add examples infrastructure}
\usage{
addExamples()
}
\value{
\code{\link[base]{character}}. Path to example directory.
}
\description{
Add examples infrastructure.
}
\details{
TODO
}
\examples{
## TODO
}

