% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addReadme.default}
\alias{addReadme.default}
\title{Add README Information}
\usage{
\method{addReadme}{default}(what, overwrite = FALSE,
  github_user = getOption(as.package(".")$package)$github_user, ...)
}
\arguments{
\item{what}{\code{ANY}.}

\item{overwrite}{\code{\link[base]{logical}}}

\item{github_user}{\code{\link[base]{character}}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{TRUE/FALSE}.
}
\description{
See generic: \code{\link[devops]{addTravisCi}}
Method for: \code{Readme.Default}
}
\details{
TODO
}
\examples{
## TODO
}

