% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addReadme}
\alias{addReadme}
\title{Add README Information}
\usage{
addReadme(what = NULL, ...)
}
\arguments{
\item{what}{A class instance.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See respective methods.
}
\description{
Adds README information to files \code{vignettes/README.Rmd} and/or
\code{README.md}.
}
\details{
TODO
}
\examples{
## TODO
}

