% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addRepositoryToRprofile}
\alias{addRepositoryToRprofile}
\title{Add repository to .Rprofile file}
\usage{
addRepositoryToRprofile(repo = c(CRAN = "http://cran.r-project.org"))
}
\arguments{
\item{repo}{\code{\link[base]{character}}.
Repository to add as named character vector.}
}
\value{
\code{TRUE}.
}
\description{
Add repository to .Rprofile file. Important for dependency checks.
}
\details{
TODO
}
\examples{
## TODO
}

