% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addToTravisCi.list}
\alias{addToTravisCi.list}
\title{Add travisCI information}
\usage{
\method{addToTravisCi}{list}(what)
}
\arguments{
\item{what}{\code{list}.}
}
\value{
\code{TRUE}.
}
\description{
See generic: \code{\link[devops]{addToTravis}}
Method for: \code{list}
}
\details{
TODO
}
\examples{
## TODO
}

