% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addToTravisCi}
\alias{addToTravisCi}
\title{Add travisCI information}
\usage{
addToTravisCi(what)
}
\arguments{
\item{what}{A class instance.}
}
\value{
See respective methods.
}
\description{
Adds information to file \code{.travis.yml}.
}
\details{
TODO
}
\examples{
## TODO
}
\seealso{
\url{http://docs.travis-ci.com/user/languages/r/}
}

