% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addToTravisCi.TravisCi}
\alias{addToTravisCi.TravisCi}
\title{Add travisCI information}
\usage{
\method{addToTravisCi}{TravisCi}(what)
}
\arguments{
\item{what}{\code{TravisCi}.}
}
\value{
\code{TRUE}.
}
\description{
See generic: \code{\link[devops]{addToTravisCi}}
Method for: \code{TravisCi}
}
\details{
TODO
}
\examples{
## TODO
}

