% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisCi.list}
\alias{addTravisCi.list}
\title{Add travisCI information}
\usage{
\method{addTravisCi}{list}(what, ...)
}
\arguments{
\item{what}{\code{list}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{TRUE}.
}
\description{
See generic: \code{\link[devops]{addTravisCi}}
Method for: \code{list}
}
\details{
TODO
}
\examples{
## TODO
}

