% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisCi}
\alias{addTravisCi}
\title{Add travisCI information}
\usage{
addTravisCi(what, ...)
}
\arguments{
\item{what}{A class instance.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See respective methods.
}
\description{
Adds information to file \code{.travis.yml}.
}
\details{
TODO
}
\examples{
## TODO
}
\seealso{
\url{http://docs.travis-ci.com/user/languages/r/}
}

