% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisCi.TravisCi.Default}
\alias{addTravisCi.TravisCi.Default}
\title{Add travisCI information}
\usage{
\method{addTravisCi}{TravisCi.Default}(what, ...)
}
\arguments{
\item{what}{\code{TravisCi.Default}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{TRUE}.
}
\description{
See generic: \code{\link[devops]{addTravisCi}}
Method for: \code{TravisCi.Default}
}
\details{
TODO
}
\examples{
## TODO
}

