% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisCi.TravisCi}
\alias{addTravisCi.TravisCi}
\title{Add travisCI information}
\usage{
\method{addTravisCi}{TravisCi}(what, path = ".travis.yml", save = TRUE, ...)
}
\arguments{
\item{what}{\code{TravisCi}.}

\item{path}{\code{\link[base]{character}}.}

\item{save}{\code{\link[base]{logical}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link[base]{character}}.
}
\description{
See generic: \code{\link[devops]{addTravisCi}}
Method for: \code{TravisCi}
}
\details{
TODO
}
\examples{
## TODO
}

