% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisInfo}
\alias{addTravisInfo}
\title{Add travisCI info to README}
\usage{
addTravisInfo(repo = "rappster/devops")
}
\arguments{
\item{repo}{\code{\link[base]{character}}.
GitHub repository.}
}
\description{
TODO
}

