% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisInfoToYaml}
\alias{addTravisInfoToYaml}
\title{Add travisCI information to YAML file}
\usage{
addTravisInfoToYaml()
}
\value{
\code{TRUE}.
}
\description{
Add travisCI information to YAML file.
}
\details{
TODO
}
\examples{
## TODO
}

