% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devops-package.r
\docType{package}
\name{devops-package}
\alias{devops}
\alias{devops-package}
\title{DevOps functionality}
\description{
Functionality that supports systematic development or what is
largely understood by the term DevOps.
}
\details{
Details
}
\section{Classes and constructors}{


\itemize{
   \item{\code{\link[devops]{devops}}: }{

     TODO
   }
}
}

\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[devops]{devops}}: }{

     TODO
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state.}
}

\section{Aknowledgements}{

This package was inspired by the following resources:

\itemize{
   \item{Package \code{devtools}: } {

     \url{https://github.com/hadley/devtools/}
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/devops}
}
\seealso{
See \code{vignette("introduction", package = "devops")}.
}

