% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_interface_classes.R
\docType{data}
\name{IReadme}
\alias{IReadme}
\title{Interface Class for README Information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IReadme
}
\description{
Defines the interface for data format information.
}
\section{Details on interface term}{

The term \emph{interface} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}

\section{Getters/setters}{


\itemize{
 \item{\code{getPath}} {

   Get file path
 }
 \item{\code{setPath}} {

   Set file path
 }
 \item{\code{getCache}} {

   Get cached content
 }
 \item{\code{setCache}} {

   Set cached content
 }
 \item{\code{getContent}} {

   Get stored content
 }
 \item{\code{setContent}} {

   Set stored content
 }
}
}

\section{Public methods}{


\itemize{
 \item{read} {
   Read README information
 }
 \item{write} {
   Write README information
 }
}
}

\section{Known interface implementations}{


\itemize{
 \item{\code{\link[devops]{Readme}}}
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/devops}
}
\keyword{datasets}

