% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{methodNotImplemented.R6}
\alias{methodNotImplemented.R6}
\title{Stop on interface call}
\usage{
\method{methodNotImplemented}{R6}(inst, cl = sys.call(-2), ...)
}
\arguments{
\item{inst}{An instance of class \code{\link[R6]{R6}}.}

\item{cl}{An instance of class \code{\link[base]{expression}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
An error informing the user that the method he called
 has not yet been implemented.
}
\description{
See generic: \code{\link[devops]{methodNotImplemented}}.

Method for instances of class \code{\link[R6]{R6}}
}
\details{
TODO
}
\examples{
## TODO
}

