% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_base_classes.R
\docType{data}
\name{Readme}
\alias{Readme}
\title{Superclass for README Information}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Readme
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{format}}{\code{\link{list}}
Data format information}
}}
\section{Getters/setters}{


\itemize{
 \item{See interface} {
   \code{\link[devops]{IReadme}}
 }
}
}

\section{Public methods}{


\itemize{
 \item{See interface} {
   \code{\link[devops]{IReadme}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/devops}
}
\keyword{datasets}

