% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{addTravisInfo}
\alias{addTravisInfo}
\title{Add travisCI info to README}
\usage{
addTravisInfo(github_user = "rappster")
}
\arguments{
\item{github_user}{\code{\link[base]{character}}.
GitHub user.}
}
\description{
TODO
}

